embedding_server/pyproject.toml
[tool.poetry]
name = "public_api_server"
version = "0.1.0"
description = "Embedding server"
authors = ["Walter Jin <jinshuhaicc@gmail.com>"]
license = "MIT License"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
pandas = "^2.2.3"
numpy = "^1.23.0"
uvicorn = "^0.32.0"
pydantic = "^2.9.2"
jinja2 = "^3.1.4"
#oci = "^2.138.0"
#oracledb = "^2.5.0"
#opencv-python = "^4.10.0.84"
pillow = "^11.0.0"
imutils = "^0.5.4"
#pydantic-settings = "^2.6.1"
asyncio = "^3.4.3"
#pytest = "^8.3.3"
#pytest-asyncio = "^0.24.0"
#insightface = "^0.7.3"
#onnxruntime = "^1.20.0"
#ultralytics = "^8.3.29"
pyyaml = "^6.0.2"
paddleocr = "^2.9.1"
paddlepaddle = "^2.6.2"
setuptools = "^75.5.0"
#opensearch-py = "^2.7.1"
#transformers = "^4.44.2"
#timm = "^1.0.11"
#einops = "^0.8.0"
clip = "^0.2.0"
python-decouple = "^3.8"
flagembedding = "^1.3.2"

#python-multipart = "^0.0.17"
#pyflakes = "^3.2.0"
#flagembedding = "^1.3.2"
#clip = "^0.2.0"
#torch = "^2.5.1"
insightface = "^0.7.3"
opencv-python = "^4.10.0.84"
onnxruntime = "^1.20.1"




[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
